N21-1904 Fix media board ghost line title (#3222)
